+++ src/include/setup.hpp
@@ -53,9 +53,9 @@ class Setup {
 
   static int64_t getMaxMemory();
 
-  static int64_t getOeisUpdateInterval();
+  static int64_t getGitHubUpdateInterval();
 
-  static int64_t getProgramsUpdateInterval();
+  static int64_t getOeisUpdateInterval();
 
   static int64_t getMaxLocalProgramAgeInDays();
 
@@ -74,11 +74,10 @@ class Setup {
   static void runWizard();
 
  private:
-  static constexpr int64_t UNDEFINED_INT = -2;                // cannot use -1
+  static constexpr int64_t UNDEFINED_INT = -2;                  // cannot use -1
+  static constexpr int64_t DEFAULT_GITHUB_UPDATE_INTERVAL = 1;  // 1 day default
   static constexpr int64_t DEFAULT_OEIS_UPDATE_INTERVAL = 7;  // 1 week default
-  static constexpr int64_t DEFAULT_PROGRAMS_UPDATE_INTERVAL =
-      1;                                                  // 1 day default
-  static constexpr int64_t DEFAULT_MAX_PROGRAM_AGE = 14;  // 2 weeks default
+  static constexpr int64_t DEFAULT_MAX_PROGRAM_AGE = 14;      // 2 weeks default
   static constexpr int64_t DEFAULT_MAX_PHYSICAL_MEMORY = 1024;  // 1 GB
 
   static std::string LODA_HOME;
@@ -90,8 +89,8 @@ class Setup {
   static bool PRINTED_MEMORY_WARNING;
   static int64_t MINING_MODE;
   static int64_t MAX_MEMORY;
+  static int64_t GITHUB_UPDATE_INTERVAL;
   static int64_t OEIS_UPDATE_INTERVAL;
-  static int64_t PROGRAMS_UPDATE_INTERVAL;
   static int64_t MAX_PROGRAM_AGE;
   static int64_t MAX_INSTANCES;